// Web/Old/authority/src/app/services/models/sub-system/SubSystemDTO.ts
export interface SubSystemDTO {
    /**
     * 唯一标识
     */
    ID: string;
    /**
     * 名称
     */
    Name: string;
    /**
     * 代码
     */
    Code: string;
    /**
     * 样式
     */
    Style: string;
    /**
     * 是否显示
     */
    Display: boolean;
    /**
     * 携带数据
     */
    Data: string;
}